4/20 Brian
Added validation to the delete word form such that user cannot delete a word if there are less than 10 words in the DB

+++ backend/controllers/wordController.js
@@ -33,8 +33,16 @@ const delWord = async (req, res) => {
     //     throw new Error('Something went wrong');
     // }
 
-    const deletedWord = await Word.deleteOne({text: req.body.text});
-    res.status(200).json(deletedWord);
+    Word.countDocuments({}, async(err,count) => {
+        if (count >= 10){
+            const deletedWord = await Word.deleteOne({text: req.body.text});
+            res.status(200).json(deletedWord);
+        } else{
+            res.send('less 10');
+        }
+
+    });
+
 };
 
 module.exports = {
